Move service url method into controller utils
