+++ src/main/app/controller/controller.utils.ts
@@ -1,3 +1,4 @@
+import config from 'config';
 import { Response } from 'express';
 
 import { endIdamSessionUrl } from '../auth/user/oidc';
@@ -5,7 +6,6 @@ import { CaseWithId } from '../case/case';
 import { YesOrNo } from '../case/definition';
 
 import { AppRequest } from './AppRequest';
-import { getServiceUrl } from './url';
 
 export const needsToExplainDelay = (userCase: Partial<CaseWithId>): boolean => {
   return (
@@ -13,6 +13,17 @@ export const needsToExplainDelay = (userCase: Partial<CaseWithId>): boolean => {
     Boolean(userCase.applicant1FinalOrderLateExplanation) ||
     Boolean(userCase.applicant2FinalOrderLateExplanation)
   );
+};
+
+export const getServiceOrigin = (req: AppRequest, res: Response): string => {
+  const protocol = req.app.locals.developmentMode ? 'http://' : 'https://';
+  const port = req.app.locals.developmentMode ? `:${config.get('port')}` : '';
+
+  return `${protocol}${res.locals.host}${port}`;
+};
+
+export const getServiceUrl = (req: AppRequest, res: Response, path: string): string => {
+  return `${getServiceOrigin(req, res)}${path}`;
 };
 
 export const destroySessionAndRedirectToSignOut = (
